fix import packages (#3)
* bitcoin API

* bcash API

* bitcoin API

* litecoin API

* ethereum API

* xrp API

* xlm API

* return pointer

* vet API

* ledger API

* fixed import packages

+++ model/response/btc/btc_tx.go
@@ -18,7 +18,7 @@ type Tx struct {
 	 * @type {number}
 	 * @memberof BtcTx
 	 */
-	Fee uint32
+	Fee uint64
 	/**
 	 *
 	 * @type {number}
